+++ qmk_keyboards/lib/tap_dance.c
@@ -13,10 +13,9 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include QMK_KEYBOARD_H
 #include "tap_dance.h"
 
-#include "lib/custom_config.h"
+#include "custom_config.h"
 
 static void on_tap_dance_finished(qk_tap_dance_state_t *state, tap_dance_data_t *data);
 static void on_tap_dance_reset(qk_tap_dance_state_t *state, tap_dance_data_t *data);
@@ -33,7 +32,7 @@ uint16_t tap_dance_get_tapping_term(uint16_t keycode, keyrecord_t *record) {
 }
 
 void tap_dance_actions_init() {
-  for (int i = 0; i < TAP_DANCE_ENTRIES; i++) {
+  for (uint8_t i = 0; i < TAP_DANCE_ENTRIES; i++) {
     tap_dance_datas[i].index = i;
     tap_dance_actions[i].fn.on_dance_finished = (void (*)(qk_tap_dance_state_t *, void *))on_tap_dance_finished;
     tap_dance_actions[i].fn.on_reset = (void (*)(qk_tap_dance_state_t *, void *))on_tap_dance_reset;
